+++ gismap/lab/egomap.py
@@ -27,6 +27,23 @@ class EgoMap(LabMap):
     ...     a.name for a in dang.authors.values() if len(a.name.split()) < 3
     ... )
     ['Bruno Kauffmann', 'Diego Perino', 'Dohy Hong', 'Fabien Mathieu', 'François Baccelli',...]
+
+    To add publications, one can use the :meth:`~gismap.lab.labmap.LabMap.add_publication` method:
+
+    >>> dang.add_publication(
+    ...     title="A new paper",
+    ...     authors=[dang.star, "Fabien Mathieu", "Alice Smith"],
+    ...     venue="Journal of Testing",
+    ... )
+    >>> str(dang.select_publications(lambda p: "Testing" in p.venue)[0])
+    'A new paper, by The-Dang Huynh, Fabien Mathieu, and Alice Smith. In Journal of Testing [unpublished], 2026.'
+
+    To remove publications, one can use the :meth:`~gismap.lab.labmap.LabMap.del_publication` method:
+
+    >>> dang.del_publication("A new paper", confirm=False)
+    >>> dang.select_publications(lambda p: "Testing" in p.venue)
+    []
+
     """
 
     def __init__(self, star, *args, **kwargs):
@@ -36,9 +53,19 @@ class EgoMap(LabMap):
         star.metadata.group = "star"
         self.star = star
         super().__init__(*args, **kwargs)
+        if self.name is None:
+            self.name = star.name
 
     def _author_iterator(self):
         yield self.star
+
+    def __repr__(self):
+        return f"EgoMap('{self.name}')"
+
+    def __str__(self):
+        n_auth = len(self.authors) if self.authors else "?"
+        n_pub = len(self.publications) if self.publications else "?"
+        return f"EgoMap of {self.name} ({n_auth} authors, {n_pub} publications)"
 
     def build(self, **kwargs):
         """